Reorganize ui scripts and prefabs
