+++ Assets/Boss/Scripts/BossController.cs
@@ -1,4 +1,5 @@
 using PseudoSummon.Audio;
+using PseudoSummon.UI;
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
