+++ Assets/UI/BoundsUI/BoundsUI.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 5f088d0e0a0d87a43ae142b222c7f3fe
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
