+++ Assets/UI/Core.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 637b7f4552a94ae49a86ac1edef42b47
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
