+++ Assets/UI/CountdownUI.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 0c8f2b6e71e1b5c428062535c37c0c92
+guid: 4bc25ff4928b9254d907be5c99620756
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
