+++ Assets/UI/OptionsMenu/OptionsMenuGroup.prefab
@@ -3262,15 +3262,16 @@ GameObject:
   serializedVersion: 6
   m_Component:
   - component: {fileID: 8348136788468317128}
-  - component: {fileID: 8348136788468317129}
   - component: {fileID: 8348136788468317127}
+  - component: {fileID: 3934861269378304372}
+  - component: {fileID: 393074581825097502}
   m_Layer: 5
   m_Name: OptionsMenuGroup
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
-  m_IsActive: 0
+  m_IsActive: 1
 --- !u!224 &8348136788468317128
 RectTransform:
   m_ObjectHideFlags: 0
@@ -3297,18 +3298,6 @@ RectTransform:
   m_AnchoredPosition: {x: 0, y: 0}
   m_SizeDelta: {x: 0, y: 0}
   m_Pivot: {x: 0.5, y: 0.5}
---- !u!225 &8348136788468317129
-CanvasGroup:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 8348136788468317130}
-  m_Enabled: 1
-  m_Alpha: 1
-  m_Interactable: 1
-  m_BlocksRaycasts: 1
-  m_IgnoreParentGroups: 0
 --- !u!114 &8348136788468317127
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -3326,6 +3315,44 @@ MonoBehaviour:
   qualityDropdown: {fileID: 6231413174487995883}
   musicTrackDropdown: {fileID: 1621682460382756752}
   mixer: {fileID: 24100000, guid: 87391913a15dc6043ac77158eeb4a625, type: 2}
+--- !u!222 &3934861269378304372
+CanvasRenderer:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 8348136788468317130}
+  m_CullTransparentMesh: 1
+--- !u!114 &393074581825097502
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 8348136788468317130}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: fe87c0e1cc204ed48ad3b37840f39efc, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  m_Material: {fileID: 0}
+  m_Color: {r: 0.1254902, g: 0, b: 0.21568628, a: 0.25490198}
+  m_RaycastTarget: 1
+  m_RaycastPadding: {x: 0, y: 0, z: 0, w: 0}
+  m_Maskable: 1
+  m_OnCullStateChanged:
+    m_PersistentCalls:
+      m_Calls: []
+  m_Sprite: {fileID: 0}
+  m_Type: 0
+  m_PreserveAspect: 0
+  m_FillCenter: 1
+  m_FillMethod: 4
+  m_FillAmount: 1
+  m_FillClockwise: 1
+  m_FillOrigin: 0
+  m_UseSpriteMesh: 0
+  m_PixelsPerUnitMultiplier: 1
 --- !u!1 &8348136788601565738
 GameObject:
   m_ObjectHideFlags: 0
