+++ Assets/UI/PauseMenu.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 59cbcb67cc55cc8468e17876cca5ad3f
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
