+++ Assets/UI/PauseMenu/PauseMenu.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 33782fbcc3693774f92a55c9b2cfb53c
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
