+++ Assets/UI/PlayerHUD.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 42546c37f52218449b25c2424322d65b
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
