+++ Assets/UI/PlayerHUD/PlayerHealthDisplay.meta
