--- Assets/UI/Scripts/UI_TitleScreen.cs
@@ -1,16 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-using UnityEngine.SceneManagement;
-using TMPro;
-
-public class UI_TitleScreen : MonoBehaviour
-{
-    public void LoadScene(string name) {
-        SceneManager.LoadScene(name);
-    }
-
-    public void QuitGame() {
-        Application.Quit();
-    }
-}
