+++ Assets/UI/TitleMenu.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: d13253ffad629f643a7d4a3cb8a976bb
+guid: 2b85b4cb9bee1704aa9fd274c17c5831
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
