+++ Assets/UI/TitleMenu/MainUIGroup.prefab.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 06fe73a5b06d6ae408e08f60f9125622
+guid: d2e4a7096ae894a4eb155b91de2bcd86
 PrefabImporter:
   externalObjects: {}
   userData:
