+++ Assets/UI/TitleMenu/TitleScreenMenu.cs
@@ -1,0 +1,18 @@
+using UnityEngine;
+using UnityEngine.SceneManagement;
+
+namespace PseudoSummon.UI
+{
+    public class TitleScreenMenu : MonoBehaviour
+    {
+        public void LoadScene(string name)
+        {
+            SceneManager.LoadScene(name);
+        }
+
+        public void QuitGame()
+        {
+            Application.Quit();
+        }
+    }
+}
