+++ Assets/UI/TutorialUI.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b08859b6e59514b4da05421a2fb2951c
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
